feat: Starting byte
Before this commit, connector would connect to the server just after
connecting to the relay. This could cause a lot of problems like
timeouts from servers. So with this, after connection and
authentication, both relay and connector are waiting. When client
connects to relay, relay sends a starting byte and then starts the data
stream. The connector connects to the server upon receiving the starting
byte and it starts the data stream as well.