fix: watch conn_a until conn_b establishes or conn_a writes
Introduces watch_stream that gets a Connection and waits until the
stream either writes something or exits without messing with the queue
(uses peek). If exits or throws error, returns true which indicates that
conn b can be aborted. if returns false, it has written something and
even if it exits the queue might be useful so it continues.

tokio::select waits until either conn_b is established or watch_stream returns
true.